Change caching behavior for file:/// paths.
Ruby changed in on of the recent versions its behavior of URI for files.
They introduced URI::File which also allows hostnames for files.
This result in a different behavior for the to string method, because it returns with file:///.
To avoid this behavior and since there are no credentials for file:, it can be returned directly.
See also for reference the new URI::File:
https://github.com/ruby/ruby/commit/04883f12c8944922117482d4d446502e5d5d3413